+++ mfhdf/test/tdatainfo.c
@@ -97,7 +97,7 @@ alloc_info(t_hdf_datainfo_t *info, unsigned info_count, int32 n_dims)
             return -1;
     }
 
-    info->dimsizes = (int32 *)malloc(n_dims * sizeof(int32));
+    info->dimsizes = (int32 *)malloc((size_t)n_dims * sizeof(int32));
     if (info->dimsizes == NULL)
         return -1;
 
@@ -384,8 +384,8 @@ test_nonspecial_SDSs()
         }
 
         /* Allocate buffers for SDS' data */
-        readibuf         = (int32 *)malloc(sds1_info.n_values * sizeof(int32));
-        readibuf_swapped = (int32 *)malloc(sds1_info.n_values * sizeof(int32));
+        readibuf         = (int32 *)malloc((size_t)sds1_info.n_values * sizeof(int32));
+        readibuf_swapped = (int32 *)malloc((size_t)sds1_info.n_values * sizeof(int32));
         /* Read in this block of data */
         readlen = read(fd, (void *)readibuf, (size_t)sds1_info.lengths[0]);
         CHECK(readlen, FAIL, "DFKconvert");
@@ -413,8 +413,8 @@ test_nonspecial_SDSs()
         }
 
         /* Allocate buffers for SDS' data */
-        readfbuf         = (float32 *)malloc(sds2_info.n_values * sizeof(float32));
-        readfbuf_swapped = (float32 *)malloc(sds2_info.n_values * sizeof(float32));
+        readfbuf         = (float32 *)malloc((size_t)sds2_info.n_values * sizeof(float32));
+        readfbuf_swapped = (float32 *)malloc((size_t)sds2_info.n_values * sizeof(float32));
         /* Read in this block of data */
         readlen = read(fd, (void *)readfbuf, (size_t)sds2_info.lengths[0]);
         CHECK(readlen, FAIL, "DFKconvert");
@@ -445,8 +445,8 @@ test_nonspecial_SDSs()
         }
 
         /* Allocate buffers for SDS' data */
-        readibuf         = (int32 *)malloc(sds3_info.n_values * sizeof(int32));
-        readibuf_swapped = (int32 *)malloc(sds3_info.n_values * sizeof(int32));
+        readibuf         = (int32 *)malloc((size_t)sds3_info.n_values * sizeof(int32));
+        readibuf_swapped = (int32 *)malloc((size_t)sds3_info.n_values * sizeof(int32));
         /* Read in this block of data */
         readlen = read(fd, (void *)readibuf, (size_t)sds3_info.lengths[0]);
         CHECK(readlen, FAIL, "DFKconvert");
@@ -1092,8 +1092,8 @@ test_chunked_partial()
         }
 
         /* Allocate buffers for SDS' data */
-        readibuf         = (int32 *)malloc(sds_info.lengths[chk_num]);
-        readibuf_swapped = (int32 *)malloc(sds_info.lengths[chk_num]);
+        readibuf         = (int32 *)malloc((size_t)sds_info.lengths[chk_num]);
+        readibuf_swapped = (int32 *)malloc((size_t)sds_info.lengths[chk_num]);
         /* readibuf = (int32 *) malloc(sds_info.n_values * sizeof(int32));
          readibuf_swapped = (int32 *) malloc(sds_info.n_values * sizeof(int32));
          */
